+++ tests/typing/cases/core/as_rotation.py
@@ -4,6 +4,7 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
+import numpy as np
 import numpy.typing as npt
 from scipy.spatial.transform import Rotation
 from type_assert import assert_types
@@ -27,10 +28,10 @@ def a_representation() -> Literal['quat', 'matrix', 'rotvec', 'mrp', 'euler', 'd
 assert_types(a_rotated_transform().as_rotation(), Rotation)
 assert_types(a_rotated_transform().as_rotation(None), Rotation)
 
-assert_types(a_rotated_transform().as_rotation('quat'), npt.NDArray[float])
-assert_types(a_rotated_transform().as_rotation('matrix'), npt.NDArray[float])
-assert_types(a_rotated_transform().as_rotation('rotvec'), npt.NDArray[float])
-assert_types(a_rotated_transform().as_rotation('euler', 'xyz'), npt.NDArray[float])
+assert_types(a_rotated_transform().as_rotation('quat'), npt.NDArray[np.float64])
+assert_types(a_rotated_transform().as_rotation('matrix'), npt.NDArray[np.float64])
+assert_types(a_rotated_transform().as_rotation('rotvec'), npt.NDArray[np.float64])
+assert_types(a_rotated_transform().as_rotation('euler', 'xyz'), npt.NDArray[np.float64])
 
 # The catch-all, reached only by a representation widened to the whole union
-assert_types(a_rotated_transform().as_rotation(a_representation()), Rotation | npt.NDArray[float])
+assert_types(a_rotated_transform().as_rotation(a_representation()), Rotation | npt.NDArray[np.float64])
