+++ tests/typing/cases/plotting/lookup_table_call.py
@@ -12,5 +12,5 @@ import pyvista as pv
 assert_types(pv.LookupTable()(0.0), tuple[float, float, float, float])
 
 # A sequence, an array or a VTK array each map to an array of colors
-assert_types(pv.LookupTable()([0.0, 1.0]), npt.NDArray[float])
-assert_types(pv.LookupTable()(np.linspace(0.0, 1.0, 4)), npt.NDArray[float])
+assert_types(pv.LookupTable()([0.0, 1.0]), npt.NDArray[np.float64])
+assert_types(pv.LookupTable()(np.linspace(0.0, 1.0, 4)), npt.NDArray[np.float64])
